Extract cloud specific code from project.py
Specific code of cloud vendor is now located in multiple classes
inheriting from Project class.

2 ways for executing specific code:
  1. Classic method overloading.
  2. Code injection using a new hook mechanism.
