+++ setup.py
@@ -25,7 +25,12 @@ setup(
     author="EDB",
     author_email="edb-devops@enterprisedb.com",
     scripts=["edb-deployment"],
-    packages=["edbdeploy", "edbdeploy.spec", "edbdeploy.commands"],
+    packages=[
+        "edbdeploy",
+        "edbdeploy.spec",
+        "edbdeploy.commands",
+        "edbdeploy.projects"
+    ],
     url="https://github.com/EnterpriseDB/postgres-deployment/",
     license="BSD",
     description=dedent("""
